mem/snpsecrets: Move SnpSecrets to its own file
With the attestation report support, the SnpSecrets is now accessed from
multiple places.
