+++ src/bios.rs
@@ -9,6 +9,7 @@
 use crate::cpu::smp_prepare_bios_vmpl;
 use crate::cpu::smp_run_bios_vmpl;
 use crate::cpu::vc::*;
+use crate::mem::SnpSecrets;
 use crate::*;
 
 use core::cmp::min;
@@ -107,47 +108,6 @@ impl SnpSection {
     pub fn size_u64(&self) -> u64 {
         self.size as u64
     }
-}
-
-#[derive(Clone, Copy, Debug)]
-#[repr(C, packed)]
-struct SnpSecrets {
-    version: u32,
-    flags: u32,
-    fms: u32,
-    reserved1: [u8; 4],
-
-    gosvw: [u8; 16],
-
-    vmpck0: [u8; 32],
-    vmpck1: [u8; 32],
-    vmpck2: [u8; 32],
-    vmpck3: [u8; 32],
-
-    os_reserved: [u8; 96],
-
-    reserved2: [u8; 64],
-
-    // SVSM fields start at offset 0x140 into the secrets page
-    svsm_base: u64,
-    svsm_size: u64,
-    svsm_caa: u64,
-    svsm_max_version: u32,
-    svsm_guest_vmpl: u8,
-    reserved3: [u8; 3],
-}
-
-#[allow(dead_code)]
-impl SnpSecrets {
-    pub fn clear_vmpck0(&mut self) {
-        self.vmpck0.iter_mut().for_each(|e| *e = 0);
-    }
-
-    funcs!(svsm_base, u64);
-    funcs!(svsm_size, u64);
-    funcs!(svsm_caa, u64);
-    funcs!(svsm_max_version, u32);
-    funcs!(svsm_guest_vmpl, u8);
 }
 
 /// 96b582de-1fb2-45f7-baea-a366c55a082d
